requirements.txt
absl-py==0.5.0
alabaster==0.7.12
altgraph==0.17
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
asgiref==3.2.10
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
astropy==3.1
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.8.0
Babel==2.7.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
baidu-aip==2.2.18.0
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
bokeh==1.0.2
boto==2.49.0
boto3==1.15.2
botocore==1.18.2
Bottleneck==1.2.1
certifi==2019.9.11
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
common==0.1.2
comtypes==1.1.7
conda==4.7.12
conda-build==3.17.6
conda-package-handling==1.6.0
conda-verify==3.1.1
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.7
cssselect==1.0.3
cycler==0.10.0
Cython==0.29.14
cytoolz==0.10.0
dask==1.0.0
datashape==0.5.4
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
distributed==1.28.1
Django==3.1.1
docutils==0.15.2
docx==0.2.4
dpkt==1.9.2
easydict==1.9
entrypoints==0.3
et-xmlfile==1.0.1
eyeD3==0.8.10
fake-useragent==0.1.11
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
future==0.17.1
gast==0.2.0
gensim==3.8.3
gevent==1.4.0
glob2==0.7
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.15.0
h5py==2.8.0
HeapDict==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==0.23
incremental==17.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itchat==1.2.32
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jieba==0.42.1
Jinja2==2.10.1
jmespath==0.10.0
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==18.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.4.2
libarchive-c==2.8
libnum==1.6.0
llvmlite==0.26.0
locket==0.2.0
lxml==4.4.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.0
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.2
model-zoo==0.2.2
more-itertools==7.2.0
mp3play==0.1.15
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
multiprocess==0.70.7
mysql-connector-python==8.0.15
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.41.0
numexpr==2.6.8
numpy==1.17.2+mkl
numpydoc==0.9.1
odo==0.5.1
olefile==0.46
opencv-python==4.1.1
openpyxl==2.6.3
packaging==19.2
pandas==0.24.2
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.4
patsy==0.5.1
pefile==2019.4.18
pep8==1.7.1
pickleshare==0.7.5
Pillow==6.1.0
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
progress==1.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
Protego==0.1.15
protobuf==3.6.1
psutil==5.6.3
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
PyDispatcher==2.0.5
pyflakes==2.1.1
pygame==1.9.4
pygeoip==0.3.2
Pygments==2.4.2
PyHamcrest==1.9.0
PyInstaller==3.6
pylint==2.3.1
PyMySQL==0.9.3
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.2.2
pypng==0.0.20
PyQRCode==1.2.1
pyquery==1.4.0
pyreadline==2.1
pyrsistent==0.15.4
PySocks==1.7.1
pytest==5.1.2
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.2.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.7.3
python-docx==0.8.10
python-Levenshtein==0.12.0
python-magic==0.4.15
pytz==2019.2
PyWavelets==1.0.1
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
queuelib==1.5.0
requests==2.22.0
rope==0.14.0
ruamel-yaml==0.15.46
s3transfer==0.3.3
scapy==2.4.3
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
Scrapy==1.8.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smart-open==2.1.1
snowballstemmer==1.9.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.8
sqlparse==0.3.1
statsmodels==0.9.0
sympy==1.4
tables==3.4.4
tblib==1.4.0
tensorboard==2.0.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.10.0
torch==1.2.0
torchvision==0.4.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
Twisted==19.7.0
unicodecsv==0.14.1
urllib3==1.24.2
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
winpcapy==1.0.2
wordcloud==1.5.0
wrapt==1.11.2
wxpy==0.3.9.8
wxPython==4.1.0
xgboost==1.0.2
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
zict==1.0.0
zipp==0.6.0
zope.interface==4.6.0
